Take the api layer to clippy 0
The last 91 findings, in two passes.

Eighty-five are `ptr_offset_with_cast` and `unnecessary_cast`, applied from
clippy's own machine-applicable suggestions and self-checking: `.offset(n as
isize)` -> `.add(n)` only compiles where `n` is already `usize`, and the
casts it drops are `bool as bool` on keyset fields and `usize as usize` on
`arglens`.

The other six needed reading:

- `nvim_buf_get_extmarks`'s reversed-range swap is two `mem::swap`s.
- Two doc comments in `api/private/helpers` and `api/ui/line` documented
  constants that earlier slices deleted, and had drifted onto the next
  constant down.
- `object_to_vim_take_luaref`'s `err` really is only used to recurse:
  upstream threads it through and never reads it. The parameter is part of
  the signature both callers spell, so it stays with a note and a
  site-level allow.

`api/` is clippy 0 in every file, and so is `ex_docmd/api.rs` alongside it.
Tree 4003 -> 3912. evalverify/varsverify/auverify byte-identical, unittest
green.

+++ crates/nvim/src/nvim/api/buffer.rs
@@ -83,8 +83,8 @@ pub const NULL: *mut ::core::ffi::c_void = ::core::ptr::null_mut::<::core::ffi::
 pub const LUA_NOREF: ::core::ffi::c_int = -2 as ::core::ffi::c_int;
 pub const INTERNAL_CALL_MASK: uint64_t = (1 as ::core::ffi::c_int as uint64_t)
     << ::core::mem::size_of::<uint64_t>()
-        .wrapping_mul(8 as usize)
-        .wrapping_sub(1 as usize);
+        .wrapping_mul(8_usize)
+        .wrapping_sub(1_usize);
 pub const VIML_INTERNAL_CALL: uint64_t = INTERNAL_CALL_MASK;
 pub const LUA_INTERNAL_CALL: uint64_t = VIML_INTERNAL_CALL.wrapping_add(1 as uint64_t);
 pub const KEYSET_OPTIDX_buf_attach__on_bytes: ::core::ffi::c_int = 2 as ::core::ffi::c_int;
